ツイートの詳細表示

+++ resources/views/tweets/index.blade.php
@@ -8,8 +8,11 @@
         <div class="more">
           <span><img src="images/arrow_top.png"></span>
           <ul class="more_list">
+          <li><a href="/tweets/{{$tweet->id}}">詳細</a></li>
+          @if (Auth::check() && Auth::user()->id == $tweet->user_id)
             <li><a href="/tweets/{{$tweet->id}}/edit">編集</a></li>
             <li><a href="/tweets/{{$tweet->id}}/delete">削除</a></li>
+          @endif
           </ul>
         </div>
       @endif
